Allow download of local Support assist without the need of a GUI
Instead of using the webbrowser to download the support assist, we can now directly hit the endpoint and write the file locally.

Also moved the accept to earlier since before it was erroring out for us. Also allow checking if we already accepted it and only accepting it if we haven't accepted the EULA.