Changed --encrypted argument to work without value
Changed --encrypted argument to work without value as passing in True on command line does not work reliable.
This change will still encrypt disks but does not allow to remove it. As removing encryption on VD's will destroy any data on it, it does make sense to just recreate the VD if you want to remove encryption.